+++ src/app/layout.tsx
@@ -1,17 +1,16 @@
 import { SpeedInsights } from '@vercel/speed-insights/next';
 import { ResolvingViewport } from 'next';
-import dynamic from 'next/dynamic';
 import { cookies } from 'next/headers';
 import { ReactNode } from 'react';
 import { isRtlLang } from 'rtl-detect';
 
 import Analytics from '@/components/Analytics';
 import { DEFAULT_LANG, LOBE_LOCALE_COOKIE } from '@/const/locale';
+import PWAInstall from '@/features/PWAInstall';
 import AuthProvider from '@/layout/AuthProvider';
 import GlobalProvider from '@/layout/GlobalProvider';
 import { isMobileDevice } from '@/utils/responsive';
 
-const PWAInstall = dynamic(() => import('@/features/PWAInstall'), { ssr: false });
 const inVercel = process.env.VERCEL === '1';
 
 type RootLayoutProps = {
